+++ packages/generator-teams/src/app/GeneratorTeamsApp.ts
@@ -615,7 +615,7 @@ export class GeneratorTeamsApp extends Generator {
         // if we have added any react based components
         if (this.options.reactComponents) {
             Yotilities.addAdditionalDeps([
-                ["msteams-react-base-component", "^3.1.1"]
+                ["msteams-react-base-component", "^4.0.0-preview"]
             ], this.fs);
         }