+++ packages/generator-teams/src/app/coreFilesUpdater/CoreFilesUpdaterFactory.ts
@@ -4,13 +4,13 @@
 
 import { BaseCoreFilesUpdater } from "./BaseCoreFilesUpdater";
 import * as semver from "semver";
-import { CoreFilesUpdater_3_6 } from "./coreFilesUpdaters/CoreFilesUpdater_3_6";
+import { CoreFilesUpdater_3 } from "./coreFilesUpdaters/CoreFilesUpdater_3";
 
 export class CoreFilesUpdaterFactory {
     public static createCoreFilesUpdater(currentVersion: string): BaseCoreFilesUpdater | undefined {
-        // < 3.6.0 cannot be updated
-        if(semver.lt(currentVersion, "3.6.0")) {
-            return new CoreFilesUpdater_3_6(currentVersion);
+        // < 4.0.0 cannot be updated
+        if(semver.lt(currentVersion, "4.0.0")) {
+            return new CoreFilesUpdater_3(currentVersion);
         }
         return undefined;
     }