+++ packages/generator-teams/src/app/manifestGeneration/ManifestGeneratorFactory.ts
@@ -9,6 +9,7 @@ import { ManifestGenerator as ManifestGenerator19 } from "./manifestGenerators/g
 import { ManifestGenerator as ManifestGenerator110 } from "./manifestGenerators/generator110/ManifestGenerator";
 import { ManifestGenerator as ManifestGenerator111 } from "./manifestGenerators/generator111/ManifestGenerator";
 import { ManifestGenerator as ManifestGeneratorDevPreview } from "./manifestGenerators/generatorDevPreview/ManifestGenerator";
+import { ManifestGenerator as ManifestGeneratorM365DevPreview } from "./manifestGenerators/generatorDevPreview/ManifestGenerator";
 import { BaseManifestGenerator } from "./BaseManifestGenerator";
 
 export class ManifestGeneratorFactory {
@@ -19,33 +20,41 @@ export class ManifestGeneratorFactory {
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.8/MicrosoftTeams.schema.json",
             manifestValue: "1.8",
             default: false,
-            hide: false
+            hide: true
         },
         {
             manifestVersion: ManifestVersions.v19,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.9/MicrosoftTeams.schema.json",
             manifestValue: "1.9",
-            default: true,
-            hide: false
+            default: false,
+            hide: true
         },
         {
             manifestVersion: ManifestVersions.v110,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.10/MicrosoftTeams.schema.json",
             manifestValue: "1.10",
             default: false,
-            hide: false
+            hide: true
         },
         {
             manifestVersion: ManifestVersions.v111,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.11/MicrosoftTeams.schema.json",
             manifestValue: "1.11",
-            default: false,
+            default: true,
             hide: false
         },
         {
             manifestVersion: ManifestVersions.devPreview,
             schemaUrl: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
             manifestValue: "devPreview",
+            default: false,
+            hide: false
+        }
+        ,
+        {
+            manifestVersion: ManifestVersions.m365DevPreview,
+            schemaUrl: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
+            manifestValue: "m365DevPreview",
             default: false,
             hide: false
         }
@@ -77,6 +86,8 @@ export class ManifestGeneratorFactory {
             return new ManifestGenerator111();
         } else if (manifestVersion == ManifestVersions.devPreview) {
             return new ManifestGeneratorDevPreview();
+        } else if (manifestVersion == ManifestVersions.m365DevPreview) {
+            return new ManifestGeneratorM365DevPreview();
         } else {
             throw new Error(`Invalid manifest version: ${manifestVersion}.`);
         }