+++ packages/generator-teams/CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [*Unreleased*]- <**>
 
+### Added
+
+* Added `m365DevPreview` manifest
 ### Changes
 
 * Updated telemetry, see https://github.com/pnp/generator-teams/blob/master/docs/docs/about/telemetry.md
 * Updated the bot template files to better reflect the official Bot Framework samples
+* Prohibited re-running the generator on version earlier than 4.0
+* Removed support for earlier manifest versions than 1.11
 
 ## [*3.4.0*]- <*2021-10-28*>